backend/requirements.txt
# Requirements for backend Flask/Grok/MusicGen API
flask
flask-cors
requests
torch
audiocraft
soundfile
sounddevice
openai
anthropic
google-generative-ai
